+++ CHANGELOG.md
@@ -7,9 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2021-03-13
+
+### Added
+
+-   New option `mergeModulesRenameDefaults` to rename default exports
+
+### Breaking Changes
+
+-   New option `mergeModulesRenameDefaults` is `true` by default.
+
 ## [1.0.0] - 2021-01-30
 
 First release
 
-[unreleased]: https://github.com/krisztianb/typedoc-plugin-merge-modules/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/krisztianb/typedoc-plugin-merge-modules/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/krisztianb/typedoc-plugin-merge-modules/releases/tag/v2.0.0
 [1.0.0]: https://github.com/krisztianb/typedoc-plugin-merge-modules/releases/tag/v1.0.0